+++ base/src/test/kotlin/proguard/util/kotlin/asserter/KotlinMetadataAsserterTest.kt
@@ -8,11 +8,16 @@
 package proguard.util.kotlin.asserter
 
 import io.kotest.core.spec.style.FreeSpec
+import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
 import io.mockk.spyk
 import io.mockk.verify
 import proguard.AppView
 import proguard.Configuration
+import proguard.classfile.Clazz
+import proguard.classfile.ProgramClass
 import proguard.classfile.kotlin.KotlinClassKindMetadata
+import proguard.classfile.kotlin.KotlinMetadata
 import proguard.classfile.kotlin.visitor.KotlinMetadataVisitor
 import proguard.resources.file.ResourceFilePool
 import proguard.testutils.ClassPoolBuilder
@@ -119,6 +124,48 @@ class KotlinMetadataAsserterTest : FreeSpec({
                         ofType<KotlinClassKindMetadata>()
                     )
                 }
+            }
+        }
+    }
+
+    "Given an enum" - {
+        val (programClassPool, libraryClassPool) = ClassPoolBuilder.fromSource(
+
+            KotlinSource(
+                "Test.kt",
+                """
+                enum class Test {
+                    CENTER,
+                    BOTTOM
+                }
+                """.trimIndent()
+            )
+
+        )
+
+        "When the referencedEnumEntries are set to null" - {
+            val visitor = spyk(
+                object : KotlinMetadataVisitor {
+                    override fun visitAnyKotlinMetadata(clazz: Clazz?, kotlinMetadata: KotlinMetadata?) {}
+
+                    override fun visitKotlinClassMetadata(
+                        clazz: Clazz?,
+                        kotlinClassKindMetadata: KotlinClassKindMetadata?
+                    ) {
+                        kotlinClassKindMetadata?.referencedEnumEntries = listOf(null, null)
+                    }
+                })
+            programClassPool.classesAccept("Test") {
+                it.kotlinMetadataAccept(visitor)
+            }
+
+            "Then the KotlinMetadataAsserter should throw away the enum metadata" {
+                (programClassPool.getClass("Test") as ProgramClass).kotlinMetadata shouldNotBe null
+                val appView = AppView(programClassPool, libraryClassPool, ResourceFilePool(), null)
+                // KotlinMetadataAsserter should remove Test enum's metadata because
+                // null entries of kotlinClassKindMetadata.referencedEnumEntries violates the ClassIntegrity.
+                KotlinMetadataAsserter(Configuration()).execute(appView)
+                (programClassPool.getClass("Test") as ProgramClass).kotlinMetadata shouldBe null
             }
         }
     }